Update tableofcontents.md
As always the solution is simple, the reason the links were not working because they needed to be the actual urls of the documents that you would get from the main branch only. That way any file updates will be kept. Unless the link changes or file path, then it would need to be updated but at least the files inside the folder as long as it stays in the same spot will be updated at all times. I also organized files inside the team folders, got rid of some empty files. But I can not get rid of the empty file in Team 4's folder because they are still trying to push files into their folder and in order to create a folder there needs to be a file inside. Once Team's 4 pull request has gone through, I will make sure deleting that extra file is a part of that request.