Update PullRequestGuide.md
There were grammatical changes that needed to be made, along with updates to folder and file names throughout the document. Furthermore, I added a section at the bottom of the document, which explains what to do once changes have been proposed. For example, this takes the viewers through filling out the pull request template, along with other small but essential details.