Rename django_backend/chatbot_be/templates/document_upload.html to prototypes/django_backend/chatbot_be/templates/document_uploa
The reason I am renaming file by file into a new folder is because when I ran the git command to move the django folder into prototypes it gave me error basically saying that it can not move a folder on main into another one on main, like the base of the repository. But if I wanted to move it into a folder further within the repository then it would actually work.